rmd-20676 split tests into different classes
